A more modest fix for the missing response-header attribute.
Removed fallback __getattr__ methods from PyramidSwaggerResponse
and PyramidSwaggerRequest, updated their docstrings to note which
properties were required, and added a specific passthrough for
response headers.

Also added a workaround to allow responses with headers but no body
to comply with the upstream interface without weird workarounds in
the response-generating code, and made the exception generated in
one special case of reading response.text not get swallowed.

Finally, changed the base class of PyramidSwaggerRequest from
RequestLike to IncomingRequest, since RequestLike is a deprecated
alias of IncomingRequest.